Extract `UpdateDatasetIconButton`
